--- restricted.php
@@ -1,5 +1,0 @@
-<?php 
-	require_once dirname(__FILE__).'/inc/globals.php';
-
-	View::render('restricted');
-?>
